pyproject.toml
[project]
name = "llama-index-memleak-repro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.14.20",
    "llama-index-workflows>=2.17.1",
]
